Rename ITs according to the naming convention
