+++ dspace-api/src/test/java/org/dspace/statistics/export/processor/BitstreamEventProcessorIT.java
@@ -9,15 +9,21 @@ package org.dspace.statistics.export.processor;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
+import static org.mockito.Mockito.mock;
 
+import java.io.File;
+import java.io.FileInputStream;
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
+import javax.servlet.http.HttpServletRequest;
 
 import org.apache.commons.codec.CharEncoding;
 import org.dspace.AbstractIntegrationTestWithDatabase;
+import org.dspace.builder.BitstreamBuilder;
 import org.dspace.builder.CollectionBuilder;
 import org.dspace.builder.CommunityBuilder;
 import org.dspace.builder.ItemBuilder;
+import org.dspace.content.Bitstream;
 import org.dspace.content.Collection;
 import org.dspace.content.Community;
 import org.dspace.content.Item;
@@ -27,23 +33,22 @@ import org.junit.Before;
 import org.junit.Test;
 
 /**
- * Test class for the ItemEventProcessor
+ * Test class for the BitstreamEventProcessor
  */
-public class ItemEventProcessorTest extends AbstractIntegrationTestWithDatabase {
+public class BitstreamEventProcessorIT extends AbstractIntegrationTestWithDatabase {
 
+    private ConfigurationService configurationService = DSpaceServicesFactory.getInstance().getConfigurationService();
 
-    private final ConfigurationService configurationService
-            = DSpaceServicesFactory.getInstance().getConfigurationService();
 
     private String encodedUrl;
 
+
     @Before
-    @Override
     public void setUp() throws Exception {
         super.setUp();
         configurationService.setProperty("irus.statistics.tracker.enabled", true);
 
-        String dspaceUrl = configurationService.getProperty("dspace.ui.url");
+        String dspaceUrl = configurationService.getProperty("dspace.server.url");
         try {
             encodedUrl = URLEncoder.encode(dspaceUrl, CharEncoding.UTF_8);
         } catch (UnsupportedEncodingException e) {
@@ -56,23 +61,27 @@ public class ItemEventProcessorTest extends AbstractIntegrationTestWithDatabase 
     /**
      * Test the method that adds data based on the object types
      */
-    public void testAddObectSpecificData() throws UnsupportedEncodingException {
+    public void testAddObectSpecificData() throws Exception {
+        HttpServletRequest request = mock(HttpServletRequest.class);
+
         context.turnOffAuthorisationSystem();
         Community community = CommunityBuilder.createCommunity(context).build();
         Collection collection = CollectionBuilder.createCollection(context, community).build();
         Item item = ItemBuilder.createItem(context, collection).build();
+
+        File f = new File(testProps.get("test.bitstream").toString());
+        Bitstream bitstream = BitstreamBuilder.createBitstream(context, item, new FileInputStream(f)).build();
+
         context.restoreAuthSystemState();
 
-        String encodedHandle = URLEncoder.encode(item.getHandle(), CharEncoding.UTF_8);
+        BitstreamEventProcessor bitstreamEventProcessor = new BitstreamEventProcessor(context, request, bitstream);
 
-        ItemEventProcessor itemEventProcessor = new ItemEventProcessor(context, null, item);
-        String result = itemEventProcessor.addObjectSpecificData("existing-string", item);
+        String result = bitstreamEventProcessor.addObjectSpecificData("existing-string", bitstream);
 
         assertThat(result,
-                   is("existing-string&svc_dat=" + encodedUrl + "%2Fhandle%2F" + encodedHandle +
-                              "&rft_dat=Investigation"));
+                   is("existing-string&svc_dat=" + encodedUrl + "%2Fapi%2Fcore%2Fbitstreams%2F" + bitstream.getID()
+                              + "%2Fcontent&rft_dat=Request"));
 
     }
-
 
 }
